THEMIS_Imple

+++ simulator/ns-3.39/src/point-to-point/model/switch-node.h
@@ -161,7 +161,7 @@ public:
 	void ClearTable();
 	bool SwitchReceiveFromDevice(Ptr<NetDevice> device, Ptr<Packet> packet, CustomHeader &ch);
 	void SwitchNotifyDequeue(uint32_t ifIndex, uint32_t qIndex, Ptr<Packet> p);
-	void SetEpsilon(uint16_t epsilon) {m_epsilon = epsilon;}
+	void SetEpsilon(uint32_t epsilon) {m_epsilon = epsilon;}
 	void ConfigureBifrostPort(uint32_t inPort, uint64_t bdpBytes, uint64_t reservedBytesH, Time slot, uint32_t k);
 	void SetBifrostPortEnabled(uint32_t inPort, bool enabled);
 	// static uint32_t cnp_count;
